signals and readline lib
