+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: aben-ham <aben-ham@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/20 18:34:52 by aben-ham          #+#    #+#             */
-/*   Updated: 2022/02/22 17:20:27 by aben-ham         ###   ########.fr       */
+/*   Updated: 2022/02/22 18:50:01 by aben-ham         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,19 +19,24 @@
 # include <signal.h>
 # include <fcntl.h>
 # include <string.h>
-#include <stdio.h>
-
-#include "readline/readline.h"
-#include "readline/history.h"
+# include <stdio.h>
+# include <readline/readline.h>
+# include <readline/history.h>
 
 # include "utils.h"
 
-typedef struct sigaction t_sigaction;
+# define PROMT_STR "minishell$ "
+# define ERR_INTERNAL "Internal Error\n"
+
+typedef struct sigaction	t_sigaction;
 
 //not used yet
 typedef struct s_gvar
 {
 	int	flag;
 }	t_gvar;
+
+void	rl_replace_line(const char *text, int i);
+int		init_sigaction(void);
 
 #endif
